Rename variable in test
